requirements.txt
toml
keyboard
pywin32
importlib
importlib-metadata
pyinstaller
PyQt6
screeninfo
thefuzz
flask
requests
psycopg2-binary
